+++ generators/templates/scaffold/model.rb
@@ -1,0 +1,4 @@
+class <%= @s_class_name %> < Prime::Model 
+  timestamp_attributes
+  attribute :title
+end