+++ generators/templates/scaffold/show.rb
@@ -1,0 +1,3 @@
+class <%= @p_class_name %>ShowSection < Prime::Section
+  element :title, text: proc { model.title }
+end